Consolidate LogicalPlan tree node walking/rewriting code into one module
